refactor Remove unused imports
With the conversion with at-noder-converter, some unused require's
were added to the code (requires that do not save the required module
to a variable). They can be removed.

Note however that there are still some require's not saved into
a variable in the code which are still needed. They are mostly:

- requires of cfg beans
- requires of filters

This is because cfg beans and filters are passed to certain methods by
their string classpath, and not by object reference.

Close #1293.