feat #1363 Improved user experience for drop down widgets on touch devices
Before this commit, when selecting an item in the popup of a
drop down widget (such as a DatePicker, an AutoComplete, a SelectBox
or a MultiSelect), the focus goes back to the input element.
On touch devices, focusing an input element triggers the display of the
keyboard. This commit prevents the focus from going back to the input
element immediately after selecting an element in the popup of a drop down
widget on touch devices or when touching the dropdown icon. The behavior
is not changed on non-touch devices (the detection of touch devices is
done through aria.utils.Device.isTouch).

close #1363