remove compat: boolean param for sync requests in submitJsonRequest
Removed backward compatibility in
`aria.templates.ModuleCtrl.submitJsonRequest` method:

Since 1.6.2 the method accepted boolean param (at 4th position)
which, when set to `false`, made the request synchronous.

This was deprecated in 1.6.7 in favor of passing an object with
options, which is more versatile and apart from setting `async` flag,
can be used to override timeout and headers.

In other words, if you used

    `submitJsonRequest(.., .., .., false)`

you should use instead:

    `submitJsonRequest(.., .., .., {async: false})`

Note that synchronous requests will freeze the UI of the browser
and should not be used unless you have a valid use case for them.

(deprecated in 1.6.7 via GH-1316)