fix #1373 Use "section" statement as a container
This commit re-introduces the possibility to use the `section` statement as a container ({section {...}}...{/section}).
Since the `macro` property is not mandatory, it is also possible to provide a string argument to the statement ({section "mySectionId"}...{/section}).

This feature is disabled by default. In order to enable it, you need to set an application environment variable:
```
aria.core.AppEnvironment.setEnvironment({
    templateSettings : {
        allowSectionsAsContainers : true
    }
});
```