fix #1377 loading overlay juddering on scroll in Safari Mobile
Since overlay implementation was using fixed positioning,
the user experience was not very good in Safari Mobile
(e.g. on ipad) because the browser doesn't do repaints
of fixed-positioned elements during scrolling.

The implementation of overlay was changed to use equivalent
absolute positioning instead.

This commit is an improved version of
2aa83bb279c1f10d11fca7ee46e52869b3c4ee67
which was reverted in
9548054e21a3c611dde40b2035fb0e0e4ba405e6

The changes since original commit were:

1) fixed ScrollableBodyTest regression causing the loading overlay
not to be displayed when page was scrolled too far

2) fixed loading overlay position when element was out of viewport:
When overlay's parent element was partially out of viewport
(on its top/left), overlay's position was badly calculated
(it was positioned too much towards the top/left).

Close #1386