fix #1377 getViewportSize on tablets
`aria.utils.Dom.getViewportSize` was not reporting
proper values on tablets.

On Chrome on Android, it was reporting virtual viewport size,
not screen-visible viewport size.

On Safari on Ipad, it was not taking into account the fact
that the viewport expands on user scroll (due to address bar
shrinking).

Close #1386