fix #1344 dialog widget: clicking on the close button interferes with drag/drop
This commit fixes the following issue: on some browsers, clicking
on the close button of the dialog is randomly interpreted as the
beginning of a drag/drop operation, preventing the dialog from
being closed.

close #1392

(cherry picked from commit 00bf1553306acb54f7294584e6f485650fb3a47a)

Conflicts:
	src/aria/widgets/container/Dialog.js

Changed CloseOrDragTest.js to not call getElementsByClassName with a null
DOM element as it was not supported in aria.jsunit.TemplateTestCase in
the 1.4.17 branch.