fix #1408 Allowing the robot not to be available immediately (waits for 2s)
Sometimes, if the first test to be executed by a browser is a robot test,
it can fail with the error "The robot is not available." even if the
browser was started with selenium-java-robot because selenium-java-robot
has not yet injected the SeleniumJavaRobot object in the page.

This commit makes sure we wait up to 2s before raising the error that the
robot is not available.