fix #1411 Classic Test Runner: load extrascripts in isolated mode
This commit adds the possibility to load in the Classic Test Runner
the extrascripts exposed in attester's config, when CTR is run
in isolated mode.

Previously, the extrascripts were injected by attester only to CTR's
index page, but CTR had no possibility to propagate those scripts
into the iframe when tests were run with `#runIsolated=true` in the URL.

This commit needs https://github.com/attester/attester/pull/121
integrated in attester (to be released in attester 2.2.0)
in order for it to properly expose `before` and `after` extrascripts.

Close #1411