feat #1415 Using the ATFillCache packaging
This commit changes the format in which Aria Templates is packaged. Instead of using the previous ATMultipart format (which does not allow packages to be evaluated in a script tag), it is now using the ATFillCache format (which fills the cache when packages are directly evaluated in a script tag).

close #1415