feat Customization of widget's error messages
It is now possible for a user to customize the error messages used by the widgets.

There are 3 levels of error messages specifications, the two firsts being user configuration while the last is the hardcoded default.

This is resolved in this order of precedence:

- local: widget's configuration
- global: application's environment
- hardcoded: widget's resources

(cherry picked from commit bbdbfe4e46dd5f5b00d7ddf9e8fe7fc8d9f65199)

Conflicts:
	test/aria/widgets/form/autocomplete/AutoCompleteTestSuite.js

feat Improved the widget's default error messages customization.

The implementation and tests code is now cleaner, and the tests are more documented. A little bit of redundancy checking has been introduced in the tests just to make sure that if the order of some methods changes or if some are removed or whatever the tests will still work and be comprehensive.

Also fixed a small mistake in AutoComplete's default error messages test.

An unwanted `return` statement was remaining at the end of the `$constructor` of the test. It was not breaking anything regarding our class system implementation, but this remains incorrect.

(cherry picked from commit 82250a1d97988a40ebc95155269d17556893f510)

feat Added backward compatibility for feature #1427

Added backward compatibility for widget's default error messages customization. The non-backward change concerns the keys used inside a resource file to define the hardcoded fallback for default values.

There is no deprecation message or so on because the broken feature was not a public feature, but a commonly agreed hack.

(cherry picked from commit c2438b7c57805f2b9f95b905ccaefed49ce5b91b)