feat 1423 Allow CSSMgr to reuse existing tags
It is important for an application to decide the order
in which widgets and templates-related css rules
(coming from CSS templates) are inserted in the head tag.

In order to do that, the application can insert tags with the correct id. This commit allows to use existing tags for the css insertion.

Close 1423