feat Removing deprecated reference to Browser.environment in own implementation.
The problem is that this is still a change of another nature compared to the others: we remove it, but we have another property which is not quite equivalent but can be used as replacement.

Concretely, `environment` was considered as a synonym of `osName`, and corresponded to the latter's value. This should not be the case, but the documentation and the deprecation logging should reflect the alternative. A custom message could be added to the definition of deprecated properties.

Closes #1421