feat Added backward compatibility for feature #1427
Added backward compatibility for widget's default error messages customization. The non-backward change concerns the keys used inside a resource file to define the hardcoded fallback for default values.

There is no deprecation message or so on because the broken feature was not a public feature, but a commonly agreed hack.