feat #1454 atLoader injector script to load AT packages
atLoader is used to synchronously load Aria Templates classes that are not included
in the AT package.
It is sufficient to insert the array of classes as the innerHTML of the tag
used to load this script.
All the classes' dependencies have to be added to this array.
No comments or other sintax is allowed in it except for the array itself.

close #1454