fix #1467 empty MultiSelect focus issue when clicking on the dropdown icon
This pull request fixes the following issue: when an @aria:MultiSelect
widget is empty (i.e. when its items property is an empty array) and
when clicking on the dropdown icon to open the popup, the widget gets
the focused style and keeps it forever (even when the focus is somewhere
else).