feat #1513 Allowing dialogs to be inside a container
This commit adds a new `container` property for the `@aria:Dialog` widget.
If this property is defined, the dialog is inserted as a child of the
corresponding DOM element (which must have the `position:relative` or
`position: absolute` style), and the dialog can only be moved or resized
inside this element. If the dialog is modal, the modal mask is only above
the container, and focusing elements outside the container is still
possible.

close #1513