fix #1521 CheckBox not appearing when using the ATFlatSkin.
The skin was using a data uri to embed the image data within the background-image of the css for the skin.
This was causing problems with IE11 resulting in the failure to load the sprite for the CheckBox.
To be consistent and to ensure the bug isn't reproduced with other svg files, the skin has been changed to use gif files instead.

close #1522

(cherry picked from commit cd730ac960ee6b73a592a8323d51029b688fa22d)