fix #1520 Adding DialogMaxWidthTestCase to a test suite
test.aria.widgets.container.dialog.sizes.DialogMaxWidthTestCase was not
in any test suite, and, as a consequence, was not run in travis.
The test failed on Travis-ci but not on some Windows machines. It seems
that Travis and on some Windows machines are using different fonts to
display the dialog title which results in different dialog widths when
the width can freely follow the width of the title.