fix #1524 Empty label issue with aria:SelectBox widget
This commit fixes the following issue: when a (non-null) value of an
aria:SelectBox widget is selected and has an empty label, clicking on the
field and immediately leaving it changes the value to null. On IE11 this
also happens when clicking on the dropdown button.

(cherry picked from commit fbb07b66c9e94e86da5546eb20087aecd576f13d)

Compared to the original commit that was cherry picked, this commit also
contains extra methods in the EmptyOptionTestCase.js file, that are
missing in the TemplateTestCase.js file in the 1.4.17 branch.