fix #1074 aria.utils.Bridge now forwards url map and root map to the popup
This commit links the url map and root map of the popup window (created
with aria.utils.Bridge) to the ones of the parent window.
It also changes a bit the logic to create the urls of the framework and
the skin included in the popup in order to make the bridge work better
(especially, to make it work in the test environment).

close #1075
close #1531
PTR 08211820