fix #1535 AdaptToContentWidthTest failing after reducing the default delay in waitFor
The condition in the waitFor call in test.aria.widgets.form.autocomplete.popupWidth.AdaptToContentWidthTest
did not correctly check whether the popup was fully loaded.
Before reducing the delay of the waitFor (in da838e16a05cc6943639b55528fc07fbcd2a238e),
the test was randomly failing, but after the delay was reduced, the test
was always failing.
This commit makes the test pass in a reliable way.