fix #1535 AdaptToContentWidthTest failing after reducing the default delay in waitFor
The condition in the waitFor call in test.aria.widgets.form.autocomplete.popupWidth.AdaptToContentWidthTest
did not correctly check whether the popup was fully loaded.
Before reducing the delay of the waitFor (in da838e1),
the test was randomly failing, but after the delay was reduced, the test
was always failing.
This commit makes the test pass in a reliable way.
It fixes an issue in commit 7e6e80c54f8864f2743af7a4dc6342fee84d3f0d.