fix #1552 wrong display of ampersand in the suggestions of the autocomplete
This commit fixes the following issue:
when typing an ampersand in the autocomplete, and if some suggestions
match, suggestions are wrongly displayed with '&' replaced by the
corresponding '&amp;' entity.
Adding another letter afterwards switches back to the correct display, but
then the highlighting in the suggestion no longer works (while there is
an ampersand in the input).

PTR 10609389