feat #1556 Adding WAI-ARIA accessibility support to the DatePicker widget
This feature adds accessibility attributes on the DatePicker DOM elements
(role, aria-activedescendant, aria-selected, aria-haspopup, aria-expanded,
aria-label...) when the waiAria widget property is true.
Also, no matter what the value of waiAria is, the browser focus is now
always on the Calendar DOM element when the popup is open, instead of
being on the dropdown button.

close #1556, CR 08528144, AT-793