fix #1570 regressions caused by the dynamic zIndex popup management feature
This commit fixes regressions introduced by commit 8fed6379173c4bfbc48b2d2d962bd197d8edde9a (#1555):
if a dialog contains widgets which have a popup (such as the dropdown
list of an aria:Select, or the validation error popup of most widgets)
the popup is often displayed behind the dialog.
This is fixed by adding a new zIndexKeepOpenOrder property on popups, which,
when set to true (which is the default), makes sure the newly opened popup
is always kept in front of any popup that was opened before.
This property is set to false on all dialogs, to make sure they can be
reordered as needed.

PTR 10900519

(cherry picked from commit 9ad62dcda7e040f38f05073e4ea85b7f9c5f9e22)