feat #1575 Allowing to write Jaws automatic tests
This commit adds the aria.jsunit.JawsTestCase base class to help writing
tests to be run with the Jaws screen reader enabled.
It also adds the test.JawsTestSuite class which is intended to reference
all test cases designed to be run with Jaws enabled.
It also includes a first Jaws test, which can also be used as an example
to write further tests: test.aria.widgets.wai.datePicker.DatePickerJawsTest1

close #1575
CR 08528144
AT-1243