fix #1593 Error in robot tests using the backspace key in PhantomJS
The backspace key was not correctly sent to PhantomJS by RobotPhantomJS
and this apparently triggered a navigation in the iframe to some previous
test, resulting in a timeout.

Especially, for the Aria Templates test campaign, this commit fixes the
random issue with AdaptToContentWidthTest, when it is run in the same
PhantomJS instance as ExternalHashNavigationTest.