fix/test test.aria.utils.dragdrop.fixedElements.FixedElementTest on IE 8
This commit fixes test.aria.utils.dragdrop.fixedElements.FixedElementTest
on IE 8, when run with vbox-robot.

On IE 8, the following strange behavior happens: when the user starts to
drag and drop a selected area (with a physical mouse, such as the one
simulated by vbox-robot, not the robot-server), the execution of
script tags is temporarily delayed until the mouse is released.
As we wait for the execution of the JSON-P request from vbox-robot before
sending the command to release the mouse, the test never ends successfully.

In this test, it is a mistake to have something selected, so, adding
the unselectable attribute fixes the issue.

close #1605