fix #1644 Listening to safetap events for popup icons of DatePicker and similar widgets
In order to fix issues when trying to open the DatePicker popup (or popups
of similar widgets), we now listen to the safetap event in addition to the
click event on touch devices.

Touch devices are detected using aria.utils.Device.isTouch().
aria.touch.SafeTap is only loaded if a touch device is detected.

PTR 11250631
AT-1326