fix #1655 Jaws reads "not checked unavailable" when the checkbox is checked (and disabled)
This commit implements a work-around for the following bug in IE 11 / Jaws 16:
in some cases a disabled checked checkbox is read as "not checked unavailable"
instead of "checked unavailable". It seems to happen when aria attributes
(such as aria-disabled or aria-labelledby) are used on the input.

AT-1470
PTR 11793963

(cherry picked from commit 72533936f5e5c1cf933325093d05c15434fa1ea4)