feat #1649 Adding the titleTag and titleClassName properties on the ErrorList widget
This commit adds the titleTag and titleClassName properties on the ErrorList widget
to allow setting the tag used for the title and its associated CSS classes
in the ErrorList widget.

This is especially useful for accessibility: screen readers often provide
a way to navigate through the heading tags of the page.

AT-1453
close #1649

(cherry picked from commit 6bf4782df217e5204b8afa3c274280d8634d007e)