feat #1650 Adding the titleTag property on the Dialog widget
This commit adds the titleTag property to the Dialog widget to allow
setting the tag used for the title in the dialog.
By default, the tag used for dialogs is now h1 when waiAria is true
and span (as before) otherwise.

This is especially useful for accessibility: screen readers often provide
a way to navigate through the heading tags of the page.

close #1650
AT-1445

(cherry picked from commit 5c89ec6e39bf7f72a9db69f0a982bbac1d8a3bb9)