fix incorrect offsetHeight on IE 7 when the dialog title is before its body
This commit implements a work-around for IE 7 because of a bug triggered
by the following commit: e65ad16a9512fc409c90417390fc506bd948e3cc (#1646)
When the title element is at the beginning of the dialog dom element,
the offsetHeight of the popup is not computed correctly by IE 7
(IE 7 inexplicably adds 4 px).

To avoid this issue, on IE 7, the title element is now inserted at the end
of the dialog element, as it was done on all browsers before commit
e65ad16a9512fc409c90417390fc506bd948e3cc.