fix focus issue on IE 7 caused by navigation elements introduced in #1654
The new navigation elements introduced in commit 6f45f150ff92f89ca8023232c299d96611c17888 (#1654)
make the following tests fail on IE 7:
 - test.aria.widgets.wai.popup.dialog.modal.FirstTest
 - test.aria.widgets.wai.popup.dialog.modal.FourthTest

Those tests are blocked in _testFocusRestoration while waiting for the
opening element to be focused again. It is not obvious why adding the
navigation elements prevents the opening element from being correctly
focused again when the dialog is closed. It must be just another IE 7 bug!

The issue seems not to happen if we do not remove the navigation elements
from the DOM when the dialog is closed. However, instead of keeping those
elements in the DOM a little longer (with a non-deterministic setTimeout),
this commit simply avoids adding those elements on IE 7 which results in
the following slightly behavior on IE 7: when pressing shift-tab from
outside the viewport when a modal dialog is open, the first item of the
dialog gets focused instead of the last one.