feat #1677 WAI-ARIA - DatePicker - Reading the input date when valid
Now when the user inputs a date, be it from the calendar or when typing in, if the date is valid it is read by the screen reader following a provided format: waiAriaConfirmDateFormat (which defaults to waiAriaDateFormat or a long date format).

When the user types the text in, a debouncing algorithm is applied, which is controlled by a delay waiAriaConfirmDateDelay, defaulting to 50 ms.

References:

- AT-1509

close #1677

(cherry picked from commit fe5721dd39045770ebc9faebe393086788681fed)