fix #1694 Using mouseenter/mouseleave instead of mouseover/mouseout in widgets
This commit adds support for mouseenter/mouseleave events in aria widgets
(by simulating those events in all browsers from the mouseover/mouseout
events).
This fixes an issue which prevented the button from working on some
browsers (including Chrome) when clicking right at the top of the <span>
element (inside the larger <button> element) because that element moves
down when clicking, leading to a mouseout event on the span but there is
no mouseleave event on the larger <button> element.

PTR 11699045
AT-1422