feat #1717 Using CORS instead of Flash by default for cross-domain requests
Now that CORS is largely implemented in browsers, this commit changes the
default implementation of cross-domain requests to use CORS instead of
the Flash-based implementation.

The Flash-based implementation can still be re-enabled by calling the
aria.core.IO.updateTransports method in the following way:

  aria.core.IO.updateTransports({
     'crossDomain' : 'aria.core.transport.XDR'
  });

Note that this change is NOT backward compatible. Applications relying on
the Flash-based implementation of cross-domain requests should use the
above lines of code or migrate to CORS.

close #1717