fix #1736 Changing implementation of interface wrappers to avoid leaks
If an object for which an interface wrapper has been created is never
disposed through its $dispose method, the garbage collector can never
collect it because the corresponding interface keeps a reference to it.

For example, assuming the x.Y class impements the x.IY interface,
the following code creates a leak:

 for (var i = 0; i < 100000; i++) {
    var myObj = new x.Y();
    myObj.$interface("x.IY");
    // The leak would not appear if myObj was disposed like this:
    // myObj.$dipsose();
 }

This commit avoids creating a leak in such a case by using a WeakMap
in the interface in order to keep a reference to the underlying object
without preventing the garbage collector from collecting the underlying
object and interface wrapper if there is no other reference to them
anywhere.
(in case the browser does not support WeakMap, an equivalent alternative
implementation is used)

AT-1602
PTR 12602212