fix #1739 memory leak when calling $getElementById or $getChild on a template
Each time a call to $getElementById or $getChild is done on a template,
a new DomElementWrapper object is created and, until this commit,
the wrapper was also added to an array in the template context, with the
only purpose of calling $dispose on all wrapper elements when
the template is fully refreshed or disposed. If this does not happen
(e.g. if there are only partial refreshes), and many calls to the
$getElementById and $getChild functions are done, this creates a leak.
With this commit, wrappers are no longer stored on the template context,
it is up to the application to dispose them (if needed).