feat #1760 Disabling early normalization of default values
This commit fixes a consistency issue between fast and slow normalization
by removing early normalization of default values.
Default values are no longer normalized at the time a bean is
pre-processed. At the time a default value is used, its copy is normalized
in the context where it is used. This can change the result of
normalization in case a bean overrides the default value of some of its
parent properties and inherits from the default value of its parent (which
is no longer normalized).