fix/test #1786 fixing issue while installing "awesome" on travis-ci
It seems that the "awesome" package raises random issues while being
installed on travis-ci (cf
https://travis-ci.org/ariatemplates/ariatemplates/builds/309427769 and
https://travis-ci.org/divdavem/ariatemplates/builds/309427661
one installation is failing, the other is successful, both from the same
commit).
This commit removes the "awesome" package from .travis.yml and also
disables the test that depended on it in order to make our build more
reliable (also, that test was not always working reliably).