fix/test Making DynamicZIndexClickRobotTestCase more reliable
Starting with commit 5bebc1478658530d477d862e0cf630bc12b25a60, it seems that
DynamicZIndexClickRobotTestCase often fails.

After investigation, it seems that when clicking on the title of a dialog,
the robot can initiate a small drag and drop operation to move the dialog,
in which case getPopupNameFromElement returns "none" for a very short
time, which was not expected in the test.