fix #1797 JAWS reads dialog title when closing dropdown widget
This commit contains a workaround for an issue with the JAWS screen reader:
JAWS reads the dialog title when closing a dropdown widget that is inside
the dialog, whereas it should not (as the user did not leave the dialog).

As investigated by @ymeine:

> - the issue comes from the fact the dropdown's element is physically
> outside the dialog, the focus is moved to it, and when closing the
> dropdown the focus is put back to the dialog, making JAWS announce it
> again
> - we can't put the dropdown's element inside the dialog because of other
> limitations (possible overflow styling on parent elements)
> - the right way to do is to use the WAI-ARIA attribute aria-owns to
> restore the logical structure of elements but JAWS doesn't support it
> - the workaround is to remove the role=dialog attribute on the Dialog's
> root element as long as a dropdown logically located inside it is open,
> and restore it short after closing this dropdown

The workaround implemented in this commit unfortunately only delays the
issue because when the user moves the focus to another item inside the
dialog, JAWS will detect the role=dialog attribute that was restored on
the dialog and read the dialog title again!

PTR 14068536