fix #1809 Wrong week numbers displayed in aria:Calendar (and aria:DatePicker)
The calendar widget now uses aria.utils.Date.getWeekNumber (which is compliant
with standard rules) to compute the week number, instead of considering
that the first week of the year is the first full week.
For example, in 2019, if the first day of the week is Monday, week 1
should start on 31/12/2018 and not on 07/01/2019.

This commit extends aria.utils.Date.getWeekNumber to allow any day of the
week to be specified as the first day of the week.
If Monday is selected, rules from ISO 8601 are applied (the first week of
the year is the week containing January 4th).
Otherwise, the first week of the year is the week containing January 1st.

PTR 16422165

(cherry picked from commit ab43afe7a77907441d6fb280b9c4627c80ce4987)