fix #1813 keyCode property in keypress event changed on Firefox 66
Firefox 66 and above has changed the keyCode property in the keypress event to match the behavior of other browsers.
This change requires an update of the aria.DomEvent class in Aria Templates because it skips some normalization of the keyCode property in the case of a Gecko-based browser. With this commit, the normalization is only skipped for Gecko browsers before version 66.