fix #1826 add a default key for Bing maps
Bing maps is not supposed to be used without a key, and it even fails
displaying a proper error message in that case.
This commit adds a default invalid key for quick tests allowing usage of
the map in case a key is not provided as it should.

(cherry picked from commit accc4ac8445ccc0eefb41e9cdf2368a7e6f09457)