feat #1830: adding role property for aria:ErrorList
This feature allows to configure the role property of the aria:ErrorList
widget. It also replaces the default "alert" value by "status"
(when ariaLive is true) because it is much less buggy with Jaws 18 and IE 11.
Especially, this change fixes the bug that makes Jaws read again the error
messages when they are removed from the screen.

PTR 19168056
close #1830