fix #1838 ua-parser-js trailing comma
The new version of ua-parser-js (0.7.26) contains a trailing comma in an
array. This is an issue with some (old) browsers.
This commit adds a reformatting step (included in the JSConcat builder of
atpackager) so that the trailing comma is removed before ua-parser-js is
included in Aria Templates (to avoid any problem).