fix #1860
As explained in the issue, in some cases, we have the previouspageenabled attribute with a wrong value causing our date picker to show the previous button.
In this fix, I removed the hour parameter in the new Date() constructor, and called removeTime to ensure that the dates are being compared without taking time into consideration.
Also, I added two new test cases to ensure that the changes are working well.

Close #1860